feat: card as a component
